backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.6.0
psycopg2-binary
python-dotenv==1.0.0
requests==2.31.0
youtube-transcript-api==0.6.1
google-cloud-aiplatform==1.38.1
nltk==3.8.1
scikit-learn
PyJWT==2.8.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-generativeai
google-auth-httplib2==0.1.1



